first tentative of refactoring in packages
