--- src/main/java/io/mosfet/lm/exercise/Product.java
@@ -1,4 +1,0 @@
-package io.mosfet.lm.exercise;
-
-public interface Product {
-}
