+++ src/main/java/io/mosfet/lm/exercise/products/Product.java
@@ -1,0 +1,4 @@
+package io.mosfet.lm.exercise.products;
+
+public interface Product {
+}
