+++ src/main/java/io/mosfet/lm/exercise/shop/Bag.java
@@ -1,4 +1,4 @@
-package io.mosfet.lm.exercise;
+package io.mosfet.lm.exercise.shop;
 
 public interface Bag {
     String getProductDescriptions();
