+++ src/test/java/io/mosfet/lm/exercise/shop/ShoppingBagTest.java
@@ -1,9 +1,11 @@
-package io.mosfet.lm.exercise;
+package io.mosfet.lm.exercise.shop;
 
+import io.mosfet.lm.exercise.products.TaxFreeProduct;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class ShoppingBagTest {
