+++ src/test/java/io/mosfet/lm/exercise/shop/SimpleCheckoutTest.java
@@ -1,5 +1,6 @@
-package io.mosfet.lm.exercise;
+package io.mosfet.lm.exercise.shop;
 
+import io.mosfet.lm.exercise.products.TaxFreeProduct;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
